requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
httpx>=0.27,<1.0
pydantic>=2.7,<3.0
pyyaml>=6.0
cryptography>=43.0
python-dotenv>=1.0
